Pool more data when pattern matching.
