+++ src/Features/Core/Portable/InitializeParameter/AbstractInitializeMemberFromParameterCodeRefactoringProviderMemberCreation.cs
@@ -483,7 +483,12 @@ namespace Microsoft.CodeAnalysis.InitializeParameter
         /// appropriate field/property names based on the user's preference.
         /// </summary>
         private List<string> GetParameterWordParts(IParameterSymbol parameter)
-            => CreateWords(StringBreaker.BreakIntoWordParts(parameter.Name), parameter.Name);
+        {
+            using (var breaks = StringBreaker.BreakIntoWordParts(parameter.Name))
+            {
+                return CreateWords(breaks, parameter.Name);
+            }
+        }
 
         private List<string> CreateWords(StringBreaks wordBreaks, string name)
         {
